+++ src/core/component/create/helpers/field.ts
@@ -9,29 +9,65 @@
 import { ComponentInterface, PropOptions, ComponentField, SystemField } from 'core/component/interface';
 import { defaultWrapper, NULL } from 'core/component/create/helpers/const';
 
-export interface FieldRealInfo {
+export interface FieldInfo {
+	path: string;
 	name: string;
+	ctx: ComponentInterface;
 	type: 'prop' | 'field' | 'system' | 'computed' | 'accessor';
 }
 
 const
 	propRgxp = /Prop$/,
-	storeRgxp = /Store$/;
+	storeRgxp = /Store$/,
+	hasSeparator = /\./;
 
 /**
- * Returns an object with real component field info
+ * Returns an object with component field info by the specified path
  *
- * @param component
- * @param name
+ * @param path
+ * @param ctx
  */
-export function getFieldRealInfo(component: ComponentInterface, name: string): FieldRealInfo {
+export function getFieldInfo(path: string, ctx: ComponentInterface): FieldInfo {
+	let
+		name = path,
+		chunks;
+
+	if (hasSeparator.test(path)) {
+		chunks = path.split('.');
+
+		let
+			rootI = 0;
+
+		let
+			obj = ctx;
+
+		for (let i = 0; i < chunks.length; i++) {
+			if (!obj) {
+				break;
+			}
+
+			obj = obj[chunks[i]];
+
+			if (obj && obj.instance instanceof ComponentInterface) {
+				ctx = obj;
+				rootI = i === chunks.length - 1 ? i : i + 1;
+			}
+		}
+
+		chunks = chunks.slice(rootI);
+		path = chunks.join('.');
+		name = chunks[0];
+	}
+
 	const
 		// @ts-ignore (access)
-		{props, fields, systemFields, computed, accessors} = component.meta;
+		{props, fields, systemFields, computed, accessors} = ctx.meta;
 
 	if (propRgxp.test(name)) {
 		return {
+			path,
 			name,
+			ctx,
 			type: 'prop'
 		};
 	}
@@ -39,34 +75,44 @@ export function getFieldRealInfo(component: ComponentInterface, name: string): F
 	if (storeRgxp.test(name)) {
 		if (fields[name]) {
 			return {
+				path,
 				name,
+				ctx,
 				type: 'field'
 			};
 		}
 
 		return {
+			path,
 			name,
+			ctx,
 			type: 'system'
 		};
 	}
 
 	if (fields[name]) {
 		return {
+			path,
 			name,
+			ctx,
 			type: 'field'
 		};
 	}
 
 	if (props[name]) {
 		return {
+			path,
 			name,
+			ctx,
 			type: 'prop'
 		};
 	}
 
 	if (systemFields[name]) {
 		return {
+			path,
 			name,
+			ctx,
 			type: 'system'
 		};
 	}
@@ -75,15 +121,29 @@ export function getFieldRealInfo(component: ComponentInterface, name: string): F
 		storeName = `${name}Store`;
 
 	if (fields[storeName]) {
+		if (chunks) {
+			chunks[0] = storeName;
+			path = chunks.join('.');
+		}
+
 		return {
-			name: storeName,
+			path,
+			name,
+			ctx,
 			type: 'field'
 		};
 	}
 
 	if (systemFields[storeName]) {
+		if (chunks) {
+			chunks[0] = storeName;
+			path = chunks.join('.');
+		}
+
 		return {
-			name: storeName,
+			path,
+			name,
+			ctx,
 			type: 'system'
 		};
 	}
@@ -92,14 +152,23 @@ export function getFieldRealInfo(component: ComponentInterface, name: string): F
 		propName = `${name}Prop`;
 
 	if (props[propName]) {
+		if (chunks) {
+			chunks[0] = propName;
+			path = chunks.join('.');
+		}
+
 		return {
-			name: propName,
+			path,
+			name,
+			ctx,
 			type: 'prop'
 		};
 	}
 
 	return {
+		path,
 		name,
+		ctx,
 		type: computed[name] ? 'computed' : accessors[name] ? 'accessor' : 'system'
 	};
 }