+++ CHANGELOG.md
@@ -11,6 +11,17 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-beta.141 (2019-08-21)
+
+#### :bug: Bug Fix
+
+* Added support for `.watch` with `@system` fields
+* Fixed `Field` API with links like `$root.something`, `$parent.something`, etc.
+  
+#### :house: Internal
+
+* Renamed `core/component/create/helpers/getRealFieldInfo` -> `getFieldInfo` and improved API
+
 ## v3.0.0-beta.140 (2019-08-21)
 
 #### :rocket: New Feature